Fix for https://github.com/wso2/product-apim/issues/2944 , and refactor code to meet eslint rules
